server/requirements.txt
fastapi
uvicorn[standard]
pdfminer.six
gensim
numpy
faiss-cpu
python-multipart